requirements.txt
gradio==4.18.0
modules==1.0.0
openai==0.27.2
python-dotenv==1.0.1
Requests==2.31.0
torch==2.0.1
transformers==4.20.0
